Addressed P.R changes.
- Changed Package name
- Added App Icon
