+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.dashx",
+    package = "com.dashxdemo.rn",
 )
 
 android_resource(
     name = "res",
-    package = "com.dashx",
+    package = "com.dashxdemo.rn",
     res = "src/main/res",
 )
